solved task
