+++ src/App.tsx
@@ -44,16 +44,6 @@ export const App: React.FC = () => {
     loadTodos(setTodos, setErrorMessage);
   }, [setTodos]);
 
-  useEffect(() => {
-    if (errorMessage) {
-      const timer = setTimeout(() => {
-        setErrorMessage(ErrorMessages.NO_ERRORS);
-      }, 3000);
-
-      return () => clearTimeout(timer);
-    }
-  }, [errorMessage]);
-
   return (
     <div className="todoapp">
       <h1 className="todoapp__title">todos</h1>
