+++ src/components/Errors.tsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames';
 import { ErrorMessages } from '../enums/ErrorMessages';
-import { Dispatch, FC, SetStateAction } from 'react';
+import { Dispatch, FC, SetStateAction, useEffect } from 'react';
 
 interface ErrorsProps {
   errorMessage: ErrorMessages;
@@ -8,6 +8,16 @@ interface ErrorsProps {
 }
 
 export const Errors: FC<ErrorsProps> = ({ errorMessage, setErrorMessage }) => {
+  useEffect(() => {
+    if (errorMessage) {
+      const timer = setTimeout(() => {
+        setErrorMessage(ErrorMessages.NO_ERRORS);
+      }, 3000);
+
+      return () => clearTimeout(timer);
+    }
+  }, [errorMessage]);
+
   return (
     <div
       data-cy="ErrorNotification"
